+++ cannon/mipsevm/evm_test.go
@@ -13,27 +13,12 @@ import (
 
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/hexutil"
-	"github.com/ethereum/go-ethereum/core/state"
 	"github.com/ethereum/go-ethereum/core/vm"
 	"github.com/ethereum/go-ethereum/eth/tracers/logger"
 	"github.com/stretchr/testify/require"
 
 	"github.com/ethereum-optimism/optimism/op-chain-ops/srcmap"
 )
-
-func testContractsSetup(t require.TestingT) (*Contracts, *Addresses) {
-	contracts, err := LoadContracts()
-	require.NoError(t, err)
-
-	addrs := &Addresses{
-		MIPS:         common.Address{0: 0xff, 19: 1},
-		Oracle:       common.Address{0: 0xff, 19: 2},
-		Sender:       common.Address{0x13, 0x37},
-		FeeRecipient: common.Address{0xaa},
-	}
-
-	return contracts, addrs
-}
 
 func SourceMapTracer(t *testing.T, contracts *Contracts, addrs *Addresses) vm.EVMLogger {
 	t.Fatal("TODO(clabby): The source map tracer is disabled until source IDs have been added to foundry artifacts.")
@@ -51,61 +36,11 @@ func MarkdownTracer() vm.EVMLogger {
 	return logger.NewMarkdownLogger(&logger.Config{}, os.Stdout)
 }
 
-type MIPSEVM struct {
-	env      *vm.EVM
-	evmState *state.StateDB
-	addrs    *Addresses
-}
-
-func NewMIPSEVM(contracts *Contracts, addrs *Addresses) *MIPSEVM {
-	env, evmState := NewEVMEnv(contracts, addrs)
-	return &MIPSEVM{env, evmState, addrs}
-}
-
-func (m *MIPSEVM) SetTracer(tracer vm.EVMLogger) {
-	m.env.Config.Tracer = tracer
-}
-
-// Step is a pure function that computes the poststate from the VM state encoded in the StepWitness.
-func (m *MIPSEVM) Step(t *testing.T, stepWitness *StepWitness) []byte {
-	sender := common.Address{0x13, 0x37}
-	startingGas := uint64(30_000_000)
-
-	// we take a snapshot so we can clean up the state, and isolate the logs of this instruction run.
-	snap := m.env.StateDB.Snapshot()
-
-	if stepWitness.HasPreimage() {
-		t.Logf("reading preimage key %x at offset %d", stepWitness.PreimageKey, stepWitness.PreimageOffset)
-		poInput, err := stepWitness.EncodePreimageOracleInput()
-		require.NoError(t, err, "encode preimage oracle input")
-		_, leftOverGas, err := m.env.Call(vm.AccountRef(sender), m.addrs.Oracle, poInput, startingGas, big.NewInt(0))
-		require.NoErrorf(t, err, "evm should not fail, took %d gas", startingGas-leftOverGas)
-	}
-
-	input := stepWitness.EncodeStepInput()
-	ret, leftOverGas, err := m.env.Call(vm.AccountRef(sender), m.addrs.MIPS, input, startingGas, big.NewInt(0))
-	require.NoError(t, err, "evm should not fail")
-	require.Len(t, ret, 32, "expecting 32-byte state hash")
-	// remember state hash, to check it against state
-	postHash := common.Hash(*(*[32]byte)(ret))
-	logs := m.evmState.Logs()
-	require.Equal(t, 1, len(logs), "expecting a log with post-state")
-	evmPost := logs[0].Data
-
-	stateHash, err := StateWitness(evmPost).StateHash()
-	require.NoError(t, err, "state hash could not be computed")
-	require.Equal(t, stateHash, postHash, "logged state must be accurate")
-
-	m.env.StateDB.RevertToSnapshot(snap)
-	t.Logf("EVM step took %d gas, and returned stateHash %s", startingGas-leftOverGas, postHash)
-	return evmPost
-}
-
 func TestEVM(t *testing.T) {
 	testFiles, err := os.ReadDir("open_mips_tests/test/bin")
 	require.NoError(t, err)
 
-	contracts, addrs := testContractsSetup(t)
+	contracts, addrs := TestContractsSetup(t)
 	var tracer vm.EVMLogger // no-tracer by default, but see SourceMapTracer and MarkdownTracer
 	//tracer = SourceMapTracer(t, contracts, addrs)
 
@@ -168,7 +103,7 @@ func TestEVM(t *testing.T) {
 }
 
 func TestEVMSingleStep(t *testing.T) {
-	contracts, addrs := testContractsSetup(t)
+	contracts, addrs := TestContractsSetup(t)
 	var tracer vm.EVMLogger
 	//tracer = SourceMapTracer(t, contracts, addrs)
 
@@ -205,7 +140,7 @@ func TestEVMSingleStep(t *testing.T) {
 }
 
 func TestEVMFault(t *testing.T) {
-	contracts, addrs := testContractsSetup(t)
+	contracts, addrs := TestContractsSetup(t)
 	var tracer vm.EVMLogger // no-tracer by default, but see SourceMapTracer and MarkdownTracer
 	//tracer = SourceMapTracer(t, contracts, addrs)
 	sender := common.Address{0x13, 0x37}
@@ -253,7 +188,7 @@ func TestEVMFault(t *testing.T) {
 }
 
 func TestHelloEVM(t *testing.T) {
-	contracts, addrs := testContractsSetup(t)
+	contracts, addrs := TestContractsSetup(t)
 	var tracer vm.EVMLogger // no-tracer by default, but see SourceMapTracer and MarkdownTracer
 	//tracer = SourceMapTracer(t, contracts, addrs)
 
@@ -304,7 +239,7 @@ func TestHelloEVM(t *testing.T) {
 }
 
 func TestClaimEVM(t *testing.T) {
-	contracts, addrs := testContractsSetup(t)
+	contracts, addrs := TestContractsSetup(t)
 	var tracer vm.EVMLogger // no-tracer by default, but see SourceMapTracer and MarkdownTracer
 	//tracer = SourceMapTracer(t, contracts, addrs)
